Extend GitHub workflow to handle all pull request branches
This commit expands the GitHub workflow configuration to include all pull request branches in addition to the "main" branch. This was necessary to ensure that the workflow runs on all pull request branches and not just the "main" branch, leading to better CI/CD integration. Also, the test definitions in 'definition_test.go' were updated with a new test attribute structure to improve our test coverage. Additionally, these changes include the creation of new files 'attr2_test.go' and 'attr2.go', introducing new functionality to the 'attribs' package, as well as making updates in '