Update README and refactor definition.go
Expanded data types listed in README.md and clarified documentation in definition.go. Data types now include float32 and float64 for a more comprehensive support. Also introduced pointer handling for array and structs for better flexibility. An additional 'ignore' argument is intended to be added in the future for ignoring unknown attributes.