Add struct pointer support and modify README
Implemented support for pointers to structs, which was previously missing. Updated README.md to inform users about the necessity of a top-level object to be an object (or pointer to an object), implemented a test to ensure the functionality of this feature. Lastly, code has been cleaned and potential confusion about tags in the example has been addressed. The diff showed some unnecessary TODO comments were removed, showing completed tasks.