Phonkee dev improvements (#7)
* Add struct pointer support and modify README

Implemented support for pointers to structs, which was previously missing. Updated README.md to inform users about the necessity of a top-level object to be an object (or pointer to an object), implemented a test to ensure the functionality of this feature. Lastly, code has been cleaned and potential confusion about tags in the example has been addressed. The diff showed some unnecessary TODO comments were removed, showing completed tasks.

* Add struct pointer support and modify README

Implemented support for pointers to structs, which was previously missing. Updated README.md to inform users about the necessity of a top-level object to be an object (or pointer to an object), implemented a test to ensure the functionality of this feature. Lastly, code has been cleaned and potential confusion about tags in the example has been addressed. The diff showed some unnecessary TODO comments were removed, showing completed tasks.

* Remove invalid attrType from attr.go

The attribute type 'invalid' was removed as it was unnecessary and not being used anywhere in the program. This also included cleaning up the code to enhance its readability and maintainability. This removal doesn't affect the other attribute types like 'integer', 'string', and 'float'.

* Replace panics with parse error in value.go

The panics in the file 'value.go' were replaced with better error handling using 'NewParseError'. This modification prevents the program from stopping execution unexpectedly, providing a more graceful way of handling errors. This change improves the robustness of the parser by handling scenarios where a set value operation cannot be performed.