Updated 'definition_test.go' and 'attr.go' to support recursive structs - Issue #10. (#11)
The 'definition_test.go' was updated to use the 'attribs_test' package and to make test cases call 'attribs.New' instead of 'New'. In 'attr.go', we updated the 'inspect' function to allow for a cache, as well as added conditionals for detecting and handling recursive structures. This update improves code efficiency and extends the functionality of the package. A 'RecursiveStruct' test case was also added to 'definition_test.go' to verify the new recursion support.